requirements.txt
Django>=4.2.24
djangorestframework>=3.15.2
django-filter>=24.0
